Split uniffi types into multiple mods
